Suppress error message when failing to set net.ipv4.conf.all.src_valid_mark.
The container needs to be run with elevated privileges in order for wg-quick to set this, however this can be set on container creation instead to avoid needing to do so. wg-quick still tries to set it even if not needed and shows an error if unable to do so.